+++ global.R
@@ -51,8 +51,8 @@ if (!.va_available) {
 }
 
 APP_TITLE <- "ClinicalVariantR"
-APP_VERSION <- "0.99.3"
-ACMG_PRO_ENGINE <- "ClinicalVariantR-Prediction-v0.99.3"
+APP_VERSION <- "0.99.4"
+ACMG_PRO_ENGINE <- "ClinicalVariantR-Prediction-v0.99.4"
 ACMG_GUIDELINE_VERSION <- "ACMG/AMP 2015 + ClinGen refinements"
 
 # Maximum upload size for VCF and companion files (1 GiB)
@@ -138,6 +138,7 @@ module_files <- c(
   "R/reproducibility.R",
   "R/acmg_engine.R",
   "R/variant_rescore.R",
+  "R/parallel_pipeline.R",
   "R/vcf_stream.R",
   "R/parse_inputs.R",
   "R/vcf_validate.R",